cloud_docker_image/requirements.txt
pillow
flask
requests
onnxruntime-gpu
opencv-python
numpy==1.23.4
speedtest-cli
matplotlib
pycocotools
mxnet==1.9.1
scipy